requirements.txt
absl-py==0.9.0
appdirs==1.4.4
apturl==0.5.2
astropy==4.2
astunparse==1.6.3
atomicwrites==1.4.0
attrs==19.3.0
Automat==20.2.0
backcall==0.1.0
bcrypt==3.1.7
bleach==3.1.5
blinker==1.4
boto3==1.13.9
botocore==1.16.9
Brlapi==0.7.0
cachetools==4.1.0
cairocffi==1.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
chrome-gnome-shell==0.0.0
cliapp==1.20180812.1
click==7.1.2
cloudpickle==0.6.1
cmdtest==0.32+git
colorama==0.4.3
coloredlogs==10.0
command-not-found==0.3
constantly==15.1.0
cryptography==3.0
cupshelpers==1.0
cycler==0.10.0
dask==2.30.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.1
distro==1.5.0
distro-info===0.23ubuntu1
docutils==0.15.2
duplicity==0.8.12.0
fasteners==0.14.1
filelock==3.0.12
Flask==1.1.2
future==0.17.1
gast==0.3.3
gestures==0.2.3
gevent==1.5.0
google-auth==1.14.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==0.4.15
grpcio==1.29.0
h5py==2.10.0
httplib2==0.18.1
humanfriendly==8.2
hyperlink==19.0.0
idna==2.10
imageio==2.9.0
importlib-metadata==1.6.0
imutils==0.5.3
incremental==17.5.0
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.0
jeepney==0.4.3
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.15.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras-Preprocessing==1.1.2
keyring==21.3.0
kiwisolver==1.2.0
klein==17.10.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
libevdev==0.5
lockfile==0.12.2
louis==3.14.0
macaroonbakery==1.3.1
Mako==1.1.2
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==2.2.5
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
nbconvert==5.6.1
nbformat==5.0.6
netifaces==0.10.4
networkx==2.5
nltk==3.5
notebook==6.0.3
numpy==1.18.4
oauthlib==3.1.0
olefile==0.46
opencv-contrib-python==4.2.0.34
opencv-python==4.4.0.46
opt-einsum==3.2.1
packaging==18.0
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.7.0
pbr==5.4.5
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.2.0
pip-review==1.1.0
pluggy==0.13.0
pooch==1.3.0
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.12.0
psutil==5.7.2
ptyprocess==0.6.0
py==1.9.0
pyamg==4.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.1
pycairo==1.16.2
pycparser==2.20
pycups==2.0.1
pyerfa==1.7.1.1
Pygments==2.6.1
PyGObject==3.38.0
PyHamcrest==2.0.2
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.4.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.16.0
pytest==4.6.11
python-apt==2.1.3+ubuntu1.2
python-crfsuite==0.9.7
python-dateutil==2.8.1
python-debian==0.1.37
python-xapp==2.0.1
pytz==2020.1
pyudev==0.21.0
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
rasa-nlu==0.15.1
regex==2019.8.19
reportlab==3.5.47
requests==2.23.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rope==0.17.0
rsa==4.0
ruamel.yaml==0.15.100
s3transfer==0.3.3
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.4.1
SecretStorage==3.1.2
Send2Trash==1.5.0
setproctitle==1.1.10
SimpleITK==2.0.1
simplejson==3.17.0
six==1.15.0
sklearn-crfsuite==0.3.6
systemd-python==234
tabulate==0.8.7
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
testresources==2.0.1
threadpoolctl==2.0.0
tifffile==2020.11.26
toolz==0.11.1
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
ttystatus==0.38
Twisted==20.3.0
typing==3.7.4.1
ubuntu-advantage-tools==24.4
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.9
usb-creator==0.3.7
virtualenv==20.0.29+ds
wadllib==1.3.4
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xkit==0.0.0
zipp==1.0.0
zope.interface==5.1.0
